moves ThemeProvider to providers folder
