+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { RouterProvider } from "react-router-dom";
-import { ThemeProvider } from './components/ThemeProvider';
 import { router } from './Router';
+import { ThemeProvider } from './providers/ThemeProvider';
 
 const App = () => {
   return (
